Change exif DateTimeDigitized to DateTimeOriginal
DateTimeDigitized is the time photo was digitized so it's the same as DateTimeOriginal
for digital cameras. I'm using DateTimeOriginal now because I've had photos without
DateTimeDigitized tag and overall it's used for photos that have been scanned to separate
the time that the photo was taken and when it was scanned.